Choose better format descriptors for updating meters
The model uses the CMETER string to determine the first part of the
command to update any external meters. In all locations (IFS, wave
model, IO server) this is defined as a 25 character string. Previously
a format descriptor of A20 was used for this, meaning the command
could be truncated.

Similarly the format descriptor for the step value the meter should
be set to was limited to 5 characters, allowing 11.4 years of
continuous running before problems occur. This is not acceptable for
multiyear configurations, and therefor this has been changed to I8
to allow run times of up to ~10,000 years.

Together these changes to format descriptors result in a CLSETEV
string that is up to 40 characters wide, which happily is the length
CLSETEV was defined to have anyway.