Switch to using `pull_request_target` trigger
The `pull_request_target` trigger is required to allow PR workflows
(here `pr-check`) to run in the context of the parent repo rather than
the fork, and hence get access to the secrets defined in the parent.

This should give access to the secrets PRTK_READ_ACCESS for read access
to the Private `prtk` repo, and to the MODIFY_PR_TOKEN to give write
access to all PRs at the Organization level. However as a side effect of
using `pull_request_target`, the standard GITHUB_TOKEN _does_ get the
`pull_request: write` permission as already defined, rather than
ignoring it, and hence MODIFY_PR_TOKEN should no longer be required.

For security given the above, we only check out the base SHA rather then
the PR head, which however means `prtk` will always check against the
original MAINTAINERS.yml as it was before the PR. This however could be
regarded as a good thing anyway.